+++ src/fbc/fbcFluxes.hxx
@@ -26,7 +26,7 @@ class Fluxes
     ~Fluxes();
     void add(Flux* flux);
     Flux* get(const char* key);
-    std::string getKey(int index);
+    std::string getKey(size_t index);
     std::vector<std::string> getKeys();
     int size();
 };
